+++ CHANGELOG.md
@@ -6,7 +6,7 @@ New features
 - Calling another action from actions
 - Allow `o "div#id.class"` or `o "div.class#id"`
 - Sub component can access the app state by adding `state:` keyword to `render`
-- `App#actions`
+- `App#actions`, `App#setup`
 - `State#==`
 - `console.log`